@@ -4,9 +4,5 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import "./style.css";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+const root = document.getElementById("root");
+ReactDOM.createRoot(root).render(<App />);
